+++ src/main/java/com/example/demo/Repositories/UserTeamRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface UserTeamRepository extends JpaRepository<UserTeam, Integer> {
+public interface UserTeamRepository extends JpaRepository<UserTeam, Long> {
 
     public List<UserTeam> findAllByUserOrderByUserTeamPointsDesc(User user);
